refactor: move mapselectcategoryvalue into @repo/shared (#1535)
