+++ app/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/components/TracksSelector/hooks/UseTable/categoryConfigs.ts
@@ -1,6 +1,6 @@
 import { GROUP_ID_LABEL } from "@/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/RelatedTracksStep/components/GenomeBrowser/components/TracksSelector/components/TracksSelectionPanel/constants";
 import { CategoryConfig } from "@databiosphere/findable-ui/lib/common/categories/config/types";
-import { mapSelectCategoryValue } from "@site-config/brc-analytics/local/index/common/category/mapSelectCategoryValue";
+import { mapSelectCategoryValue } from "@repo/shared/utils/mapSelectCategoryValue";
 
 export const CATEGORY_CONFIGS: Record<string, CategoryConfig> = {
   GROUP_ID: {
