+++ packages/shared/utils/mapSelectCategoryValue.ts
